Finish moving of files and fix namespaces
