+++ bundles/org.openhab.binding.hdpowerview/src/main/java/org/openhab/binding/hdpowerview/internal/HDPowerViewWebTargets.java
@@ -27,32 +27,32 @@ import org.eclipse.jetty.client.util.StringContentProvider;
 import org.eclipse.jetty.http.HttpHeader;
 import org.eclipse.jetty.http.HttpMethod;
 import org.eclipse.jetty.http.HttpStatus;
-import org.openhab.binding.hdpowerview.internal.api.Color;
-import org.openhab.binding.hdpowerview.internal.api.HubFirmware;
-import org.openhab.binding.hdpowerview.internal.api.Scene;
-import org.openhab.binding.hdpowerview.internal.api.SceneCollection;
-import org.openhab.binding.hdpowerview.internal.api.ScheduledEvent;
-import org.openhab.binding.hdpowerview.internal.api.ShadeData;
-import org.openhab.binding.hdpowerview.internal.api.ShadePosition;
-import org.openhab.binding.hdpowerview.internal.api.SurveyData;
-import org.openhab.binding.hdpowerview.internal.api.UserData;
-import org.openhab.binding.hdpowerview.internal.api.requests.RepeaterBlinking;
-import org.openhab.binding.hdpowerview.internal.api.requests.RepeaterColor;
-import org.openhab.binding.hdpowerview.internal.api.requests.ShadeCalibrate;
-import org.openhab.binding.hdpowerview.internal.api.requests.ShadeJog;
-import org.openhab.binding.hdpowerview.internal.api.requests.ShadeMove;
-import org.openhab.binding.hdpowerview.internal.api.requests.ShadeStop;
-import org.openhab.binding.hdpowerview.internal.api.responses.FirmwareVersion;
-import org.openhab.binding.hdpowerview.internal.api.responses.Repeater;
-import org.openhab.binding.hdpowerview.internal.api.responses.RepeaterData;
-import org.openhab.binding.hdpowerview.internal.api.responses.Repeaters;
-import org.openhab.binding.hdpowerview.internal.api.responses.SceneCollections;
-import org.openhab.binding.hdpowerview.internal.api.responses.Scenes;
-import org.openhab.binding.hdpowerview.internal.api.responses.ScheduledEvents;
-import org.openhab.binding.hdpowerview.internal.api.responses.Shade;
-import org.openhab.binding.hdpowerview.internal.api.responses.Shades;
-import org.openhab.binding.hdpowerview.internal.api.responses.Survey;
-import org.openhab.binding.hdpowerview.internal.api.responses.UserDataResponse;
+import org.openhab.binding.hdpowerview.internal.dto.Color;
+import org.openhab.binding.hdpowerview.internal.dto.HubFirmware;
+import org.openhab.binding.hdpowerview.internal.dto.Scene;
+import org.openhab.binding.hdpowerview.internal.dto.SceneCollection;
+import org.openhab.binding.hdpowerview.internal.dto.ScheduledEvent;
+import org.openhab.binding.hdpowerview.internal.dto.ShadeData;
+import org.openhab.binding.hdpowerview.internal.dto.ShadePosition;
+import org.openhab.binding.hdpowerview.internal.dto.SurveyData;
+import org.openhab.binding.hdpowerview.internal.dto.UserData;
+import org.openhab.binding.hdpowerview.internal.dto.requests.RepeaterBlinking;
+import org.openhab.binding.hdpowerview.internal.dto.requests.RepeaterColor;
+import org.openhab.binding.hdpowerview.internal.dto.requests.ShadeCalibrate;
+import org.openhab.binding.hdpowerview.internal.dto.requests.ShadeJog;
+import org.openhab.binding.hdpowerview.internal.dto.requests.ShadeMove;
+import org.openhab.binding.hdpowerview.internal.dto.requests.ShadeStop;
+import org.openhab.binding.hdpowerview.internal.dto.responses.FirmwareVersion;
+import org.openhab.binding.hdpowerview.internal.dto.responses.Repeater;
+import org.openhab.binding.hdpowerview.internal.dto.responses.RepeaterData;
+import org.openhab.binding.hdpowerview.internal.dto.responses.Repeaters;
+import org.openhab.binding.hdpowerview.internal.dto.responses.SceneCollections;
+import org.openhab.binding.hdpowerview.internal.dto.responses.Scenes;
+import org.openhab.binding.hdpowerview.internal.dto.responses.ScheduledEvents;
+import org.openhab.binding.hdpowerview.internal.dto.responses.Shade;
+import org.openhab.binding.hdpowerview.internal.dto.responses.Shades;
+import org.openhab.binding.hdpowerview.internal.dto.responses.Survey;
+import org.openhab.binding.hdpowerview.internal.dto.responses.UserDataResponse;
 import org.openhab.binding.hdpowerview.internal.exceptions.HubInvalidResponseException;
 import org.openhab.binding.hdpowerview.internal.exceptions.HubMaintenanceException;
 import org.openhab.binding.hdpowerview.internal.exceptions.HubProcessingException;
